Define a rough roadmap of training modes that this repo will aim to support, and rename things accordingly.
