+++ pyproject.toml
@@ -43,7 +43,8 @@ dependencies = [
 ]
 
 [project.scripts]
-"invoke-train-lora" = "invoke_training.scripts.invoke_train_lora:main"
+"invoke-finetune-lora-sd" = "invoke_training.scripts.invoke_finetune_lora_sd:main"
+"invoke-finetune-lora-sdxl" = "invoke_training.scripts.invoke_finetune_lora_sdxl:main"
 
 [project.urls]
 "Homepage" = "https://github.com/invoke-ai/invoke-training"
