+++ src/invoke_training/training/finetune_lora/finetune_lora_sdxl.py
@@ -1,0 +1,7 @@
+from invoke_training.training.finetune_lora.finetune_lora_config import (
+    FinetuneLoRAConfig,
+)
+
+
+def run_training(config: FinetuneLoRAConfig):
+    raise NotImplementedError("finetune_lora_sdxl is not implemented.")
