+++ src/invoke_training/training/shared/datasets/image_caption_dataloader.py
@@ -1,7 +1,7 @@
 from torch.utils.data import DataLoader
 from transformers import CLIPTokenizer
 
-from invoke_training.training.lora.lora_training_config import DatasetConfig
+from invoke_training.training.finetune_lora.finetune_lora_config import DatasetConfig
 from invoke_training.training.shared.datasets.hf_dir_image_caption_reader import (
     HFDirImageCaptionReader,
 )
